[WIP] Fix build due to missing file prefs/prefs_behavior.js
This has moved to ChromeOS-specific files now, so commenting it out
for now just to fix the build, but we need to figure out whether
simply removing references to PrefsBehaviour is good enough or more
changes need to be done.

TODO: Check with Pete whether this can be migrated to using
prefs_mixin.js instead, like upstream does now.

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/79004eab4da931e20858e7de531db8c4cb5c82dd

commit 79004eab4da931e20858e7de531db8c4cb5c82dd
Author: dpapad <dpapad@chromium.org>
Date:   Fri Oct 1 07:19:22 2021 +0000

    Settings: Move prefs_behavior.js to CrOS Settings, only used there.

    Browser Settings uses prefs_mixin.js instead. CrOS Settings can't
    leverage Mixins yet, as it is using the legacy Polymer syntax.

    Bug: 1234307